buffer.c uses ssize_t cast to squash signedness warnings

+++ src/buffer.c
@@ -215,7 +215,7 @@ buffer_set_allocate_item_locked(struct buffer_set *bs)
     bs->bs_num_allocated++;
     bi->bi_allocated = true;
 
-    if (bs->bs_num_allocated > bs->bs_max_allocated)
+    if (bs->bs_num_allocated > (ssize_t)bs->bs_max_allocated)
         bs->bs_max_allocated = bs->bs_num_allocated;
 
     return bi;
@@ -257,12 +257,12 @@ buffer_set_allocate_item_from_pending(struct buffer_set *bs)
 int
 buffer_set_release_pending_alloc(struct buffer_set *bs, const size_t nitems)
 {
-    if (!bs || nitems > bs->bs_num_bufs)
+    if (!bs || (ssize_t)nitems > bs->bs_num_bufs)
         return -EINVAL;
 
     BS_LOCK(bs);
 
-    if (nitems > bs->bs_num_pndg_alloc)
+    if ((ssize_t)nitems > bs->bs_num_pndg_alloc)
     {
         BS_UNLOCK(bs);
 
@@ -284,7 +284,7 @@ buffer_set_pending_alloc(struct buffer_set *bs, const size_t nitems)
 
     BS_LOCK(bs);
 
-    if (bs->bs_num_bufs < nitems)
+    if (bs->bs_num_bufs < (ssize_t)nitems)
     {
         BS_UNLOCK(bs);
         return -ENOMEM;
